Add some basic validity-checking for compiled dictionaries
Function to confirm whether a given file looks like it could be a valid
compiled hyphenation dictionary.

Added a "magic number" to the start of the compiled file format,
as a quick check that we've been given the right kind of file. (And
as a potential version indicator in case of future revisions.)

Also made state and string accessors more robust, aiming to safely return
empty values instead of panicking if the file is broken.